fix: 여행 정보 탭 정보 리스트 컴포넌트 분리
